backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
motor==3.3.2
beanie==1.23.0
pymongo==4.6.0

# Redis
redis==5.0.1
hiredis==2.2.3

# Celery
celery==5.3.4
celery[redis]==5.3.4

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Excel export
openpyxl==3.1.2

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code quality
ruff==0.1.6
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
